==> 03_data_cleaning.ipynb <==
import base64
import datetime
import time
import json
import numpy as np
import pandas as pd
import matplotlib as mpl
import matplotlib.pyplot as plt
from pylab import rcParams
import seaborn as sns
import statsmodels as sm
from pylab import rcParams
from numpy import median
import matplotlib.dates as mdates

np.random.seed(1337)
%matplotlib inline
sns.set(font_scale=1.5)
rcParams['figure.figsize'] = 12, 8
sns.set_style('whitegrid')
sns.set_palette(sns.color_palette('muted'))

## Load the Data

tr = pd.read_csv('input/train_ver2.csv',
        dtype={
        'age': np.str,
        'antiguedad': np.str,
        'indrel_1mes': np.str,
        'conyuemp': np.str},
        parse_dates=[
        'fecha_dato',
        'fecha_alta'])
len(tr)

te = pd.read_csv('input/test_ver2.csv',
        dtype={
        'age': np.str,
        'antiguedad': np.str,
        'indrel_1mes': np.str,
        'conyuemp': np.str},
        parse_dates=[
        'fecha_dato',
        'fecha_alta'])
len(te)

for c in tr.columns[24:]:
    te[c] = np.nan

d = pd.concat([tr, te])
len(d)

del tr
del te

## measurement_date

d.rename(columns={'fecha_dato': 'measurement_date'}, inplace=True)

col_name = 'measurement_date'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## customer_id

d.rename(columns={'ncodpers': 'customer_id'}, inplace=True)

col_name = 'customer_id'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## employment_index

d.rename(columns={'ind_empleado': 'employment_index'}, inplace=True)

col_name = 'employment_index'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.employment_index = d.employment_index.map({
        'N': 'not-employee',
        'B': 'ex-employee',
        'F': 'filial',
        'A': 'active',
        'S': 'passive'})

col_name = 'employment_index'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## country_of_residence

d.rename(columns={'pais_residencia': 'country_of_residence'}, inplace=True)

col_name = 'country_of_residence'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## gender

d.rename(columns={'sexo': 'gender'}, inplace=True)

col_name = 'gender'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.gender = d.gender.map({
        'V': 'male',
        'H': 'female'})

## age

col_name = 'age'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

def fix_age(x):
    try:
        a = int(x)
        if a < 0 or a > 110:
            return np.nan
        return a
    except:
        return np.nan

d.age = d.age.apply(fix_age)

col_name = 'age'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.age.dropna().describe()

## first_contract_date

d.rename(columns={'fecha_alta': 'first_contract_date'}, inplace=True)

col_name = 'first_contract_date'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## is_new_customer

d.rename(columns={'ind_nuevo': 'is_new_customer'}, inplace=True)

col_name = 'is_new_customer'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.is_new_customer = d.is_new_customer.map({
        0.0: 'no',
        1.0: 'yes'
    })

col_name = 'is_new_customer'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## customer_for_months

d.rename(columns={'antiguedad': 'customer_for_months'}, inplace=True)

col_name = 'customer_for_months'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

def fix_customer_for_months(x):
    try:
        a = int(x)
        if a < 0 or a > 1200:
            return np.nan
        return a
    except:
        return np.nan

d.customer_for_months = d.customer_for_months.apply(fix_customer_for_months)

col_name = 'customer_for_months'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.customer_for_months.dropna().describe()

## is_primary_customer

d.rename(columns={'indrel': 'is_primary_customer'}, inplace=True)

col_name = 'is_primary_customer'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.is_primary_customer = d.is_primary_customer.map({
        1: 'yes',
        99: 'no'})

col_name = 'is_primary_customer'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## last_primary_customer_date

d.rename(columns={'ult_fec_cli_1t': 'last_primary_customer_date'}, inplace=True)

col_name = 'last_primary_customer_date'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## customer_type

d.rename(columns={'indrel_1mes': 'customer_type'}, inplace=True)

col_name = 'customer_type'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.customer_type = d.customer_type.map({
        1: 'primary',
        1.0: 'primary',
        '1': 'primary',
        '1.0': 'primary',
        2: 'co-owner',
        2.0: 'co-owner',
        '2': 'co-owner',
        '2.0': 'co-owner',
        3: 'ex-primary',
        3.0: 'ex-primary',
        '3': 'ex-primary',
        '3.0': 'ex-primary',
        4: 'ex-co-owner',
        4.0: 'ex-co-owner',
        '4': 'ex-co-owner',
        '4.0': 'ex-co-owner',
        'P': 'potential'})

col_name = 'customer_type'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## customer_relation_type

d.rename(columns={'tiprel_1mes': 'customer_relation_type'}, inplace=True)

col_name = 'customer_relation_type'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.customer_relation_type = d.customer_relation_type.map({
        'I': 'inactive',
        'A': 'active',
        'P': 'ex-customer',
        'R': 'potential'})

col_name = 'customer_relation_type'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## country_residence_bank_same

d.rename(columns={'indresi': 'country_residence_bank_same'}, inplace=True)

col_name = 'country_residence_bank_same'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.country_residence_bank_same = d.country_residence_bank_same.map({
        'S': 'yes',
        'N': 'no'})

col_name = 'country_residence_bank_same'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## is_foreigner

d.rename(columns={'indext': 'is_foreigner'}, inplace=True)

col_name = 'is_foreigner'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.is_foreigner = d.is_foreigner.map({
        'S': 'yes',
        'N': 'no'})

col_name = 'is_foreigner'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## is_spouse_of_an_employee

d.rename(columns={'conyuemp': 'is_spouse_of_an_employee'}, inplace=True)

col_name = 'is_spouse_of_an_employee'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.is_spouse_of_an_employee = d.is_spouse_of_an_employee.map({
        'S': 'yes',
        'N': 'no'})

col_name = 'is_spouse_of_an_employee'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## join_channel

d.rename(columns={'canal_entrada': 'join_channel'}, inplace=True)

col_name = 'join_channel'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## is_dead

d.rename(columns={'indfall': 'is_dead'}, inplace=True)

col_name = 'is_dead'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.is_dead = d.is_dead.map({
        'S': 'yes',
        'N': 'no'})

col_name = 'is_dead'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## address_type

d.rename(columns={'tipodom': 'address_type'}, inplace=True)

col_name = 'address_type'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.address_type = d.address_type.map({
        1.0: 'address'})

col_name = 'address_type'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## province_code

d.rename(columns={'cod_prov': 'province_code'}, inplace=True)

col_name = 'province_code'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

def fix_province_code(c):
    try:
        return 'prov_' + str(int(c))
    except:
        return np.nan

d.province_code = d.province_code.apply(fix_province_code)

col_name = 'province_code'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## province_name

d.rename(columns={'nomprov': 'province_name'}, inplace=True)

col_name = 'province_name'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

del d['province_name']

## is_active_customer

d.rename(columns={'ind_actividad_cliente': 'is_active_customer'}, inplace=True)

col_name = 'is_active_customer'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.is_active_customer = d.is_active_customer.map({
        1.0: 'yes',
        0.0: 'no'})

col_name = 'is_active_customer'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

## household_income

d.rename(columns={'renta': 'household_income'}, inplace=True)

col_name = 'household_income'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

def fix_household_income(i):
    try:
        return float(i)
    except:
        return np.nan

d.household_income = d.household_income.apply(fix_household_income)

col_name = 'household_income'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.household_income.dropna().describe()

## customer_segment

d.rename(columns={'segmento': 'customer_segment'}, inplace=True)

col_name = 'customer_segment'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.customer_segment = d.customer_segment.map({
        '02 - PARTICULARES': 'individual',
        '01 - TOP': 'vip',
        '03 - UNIVERSITARIO': 'college-graduated'})

col_name = 'customer_segment'

print '-' * 30, '\n', 'dtype:           ', d[col_name].dtype
print '-' * 30, '\n', 'np.na count:     ', d[col_name].isnull().sum()
print '-' * 30, '\n', 'np.na ratio:     ', 1.0 * d[col_name].isnull().sum() / len(d)
print '-' * 30, '\n', 'distinct values: ', len(d[col_name].value_counts())
print '-' * 30, '\n', 'top 10 values:'
print d[col_name].value_counts().head(10)
print '-' * 30, '\n', 'bottom 10 values:'
print d[col_name].value_counts().tail(10)
print '-' * 30

d.head(10).T

## Write the Data to a File

d.to_csv('tmp/clean.csv', index=False)
