==> 04_sample_for_aggregation.ipynb <==
import base64
import datetime
import time
import json
import numpy as np
import pandas as pd
import matplotlib as mpl
import matplotlib.pyplot as plt
from pylab import rcParams
import seaborn as sns
import statsmodels as sm
from pylab import rcParams
from numpy import median
import matplotlib.dates as mdates
from random import shuffle

np.random.seed(1337)
%matplotlib inline
sns.set(font_scale=1.5)
rcParams['figure.figsize'] = 12, 8
sns.set_style('whitegrid')
sns.set_palette(sns.color_palette('muted'))

d = pd.read_csv('tmp/clean.csv')
len(d)

cids = d.customer_id.unique()
len(cids)

shuffle(cids)

cids_100 = cids[:100]
cids_1000 = cids[:1000]
cids_10000 = cids[:10000]
cids_30000 = cids[:30000]
cids_100000 = cids[:100000]
cids_300000 = cids[:300000]

d_100 = d[d.customer_id.isin(cids_100)]
d_1000 = d[d.customer_id.isin(cids_1000)]
d_10000 = d[d.customer_id.isin(cids_10000)]
d_30000 = d[d.customer_id.isin(cids_30000)]
d_100000 = d[d.customer_id.isin(cids_100000)]
d_300000 = d[d.customer_id.isin(cids_300000)]

d_100.to_csv('tmp/clean_sample_100.csv', index=False)
d_1000.to_csv('tmp/clean_sample_1000.csv', index=False)
d_10000.to_csv('tmp/clean_sample_10000.csv', index=False)
d_30000.to_csv('tmp/clean_sample_30000.csv', index=False)
d_100000.to_csv('tmp/clean_sample_100000.csv', index=False)
d_300000.to_csv('tmp/clean_sample_300000.csv', index=False)

print '\n'.join(d.columns[23:].tolist())

for i in range(10):
    cids_i = cids[(i*100000):((i+1)*100000)]
    d_i = d[d.customer_id.isin(cids_i)]
    d_i.to_csv('tmp/clean_part_{0}.csv'.format(i), index=False)
