==> 08_data_selection.ipynb <==
import base64
import datetime
import time
import json
import numpy as np
import pandas as pd
import matplotlib as mpl
import matplotlib.pyplot as plt
from pylab import rcParams
import seaborn as sns
import statsmodels as sm
from pylab import rcParams
from numpy import median
import matplotlib.dates as mdates

np.random.seed(1337)
%matplotlib inline
sns.set(font_scale=1.5)
rcParams['figure.figsize'] = 12, 8
sns.set_style('whitegrid')
sns.set_palette(sns.color_palette('muted'))

import gc

products = pd.read_csv('tmp/raw.csv', nrows=1).columns.tolist()[23:]
for p in products:
    print p

d_ho = pd.read_csv('tmp/raw.csv', usecols=['measurement_date', 'customer_id'] + products)
d_ho = d_ho[d_ho.measurement_date != '2016-06-28']
len(d_ho)

dg = d_ho.groupby(['customer_id'])
del d_ho
gc.collect()

ids = {}
for p in products:
    ids[p + "_pos"] = []
    ids[p + "_neg"] = []

i = 0
for k, g in dg:
    i += 1
    if i % 1000 == 0:
        break
    if i % 10000 == 0:
        print i / 10000
        sc.collect()
    gs = g.sort_values(by='measurement_date')
    for p in products:
        if len(gs[p].dropna().value_counts()) > 1 and gs[p].dropna().tolist()[0] == 0.0:
            ids[p + "_pos"].append(k)
        else:
            ids[p + "_neg"].append(k)

for p in products:
    print p + '_pos: ', len(ids[p + "_pos"])
    print p + '_neg: ', len(ids[p + "_neg"])

for p in products:
    pd.DataFrame({'customer_id': ids[p + '_pos']}).to_csv('tmp/pos_ids_' + p + '.csv', index=False)
    pd.DataFrame({'customer_id': ids[p + '_neg']}).to_csv('tmp/neg_ids_' + p + '.csv', index=False)

18:17

del dg

dg = pd.read_csv('tmp/clean_ho.csv').groupby(['customer_id'])
len(dg)

ids = {}
for p in products:
    ids[p + "_pos"] = []
    ids[p + "_neg"] = []
gc.collect()
i = 0
for k, g in dg:
    i += 1
    if i % 10000 == 0:
        print i / 10000
    gs = g.sort_values(by='measurement_date')
    for p in products:
        if len(gs[p].dropna().value_counts()) > 1 and gs[p].dropna().tolist()[0] == 0.0:
            ids[p + "_pos"].append(k)
        else:
            ids[p + "_neg"].append(k)
for p in products:
    print p + '_pos: ', len(ids[p + "_pos"])
    print p + '_neg: ', len(ids[p + "_neg"])
for p in products:
    pd.DataFrame({'customer_id': ids[p + '_pos']}).to_csv('tmp/clean_ho_' + p + '_ids_pos.csv', index=False)
    pd.DataFrame({'customer_id': ids[p + '_neg']}).to_csv('tmp/clean_ho_' + p + '_ids_neg.csv', index=False)

del dg
gc.collect()

d = pd.read_csv('tmp/clean_ho.csv')

d.head()

for p in products:
    d.loc[d.measurement_date == '2016-05-28', p] = np.nan

d.to_csv('tmp/clean_ho.csv', index=False)

import sklearn.utils

products1 = '''ind_recibo_ult1
ind_cco_fin_ult1
ind_nom_pens_ult1
ind_nomina_ult1
ind_cno_fin_ult1
ind_tjcr_fin_ult1
ind_ecue_fin_ult1
ind_dela_fin_ult1'''.split('\n')
products2 = '''ind_reca_fin_ult1
ind_ctma_fin_ult1
ind_valo_fin_ult1
ind_fond_fin_ult1
ind_deco_fin_ult1
ind_ctop_fin_ult1
ind_ctpp_fin_ult1
ind_plan_fin_ult1'''.split('\n')
products3 = '''ind_ctju_fin_ult1
ind_deme_fin_ult1
ind_cder_fin_ult1
ind_pres_fin_ult1
ind_hip_fin_ult1
ind_viv_fin_ult1
ind_aval_fin_ult1
ind_ahor_fin_ult1'''.split('\n')



d = pd.read_csv('tmp/raw.csv')

for p in products3:
    ids = pd.read_csv('tmp/train_ids_{0}.csv'.format(p))
    print p, len(ids)
    e = pd.merge(ids, d, on='customer_id', how='inner')
    e.to_csv('tmp/raw_train_{0}.csv'.format(p), index=False)
    del e
    del ids

d = pd.read_csv('tmp/raw.csv')
cids = d.customer_id.unique()
len(cids)

for i in range(10):
    cids_i = cids[(i*100000):((i+1)*100000)]
    d_i = d[d.customer_id.isin(cids_i)]
    d_i.to_csv('tmp/raw_part{0}.csv'.format(i), index=False)

d = pd.read_csv('tmp/raw_ho.csv')
cids = d.customer_id.unique()
len(cids)

for i in range(10):
    cids_i = cids[(i*100000):((i+1)*100000)]
    d_i = d[d.customer_id.isin(cids_i)]
    d_i.to_csv('tmp/raw_ho_part{0}.csv'.format(i), index=False)

products = '''ind_ahor_fin_ult1
ind_aval_fin_ult1
ind_cco_fin_ult1
ind_cder_fin_ult1
ind_cno_fin_ult1
ind_ctju_fin_ult1
ind_ctma_fin_ult1
ind_ctop_fin_ult1
ind_ctpp_fin_ult1
ind_deco_fin_ult1
ind_deme_fin_ult1
ind_dela_fin_ult1
ind_ecue_fin_ult1
ind_fond_fin_ult1
ind_hip_fin_ult1
ind_plan_fin_ult1
ind_pres_fin_ult1
ind_reca_fin_ult1
ind_tjcr_fin_ult1
ind_valo_fin_ult1
ind_viv_fin_ult1
ind_nomina_ult1
ind_nom_pens_ult1
ind_recibo_ult1'''.split('\n')

d = pd.read_csv('tmp/ho_test_labels.csv').sort_values(['customer_id', 'measurement_date'])
d[:20]

len(d)

dg = pd.read_csv('tmp/ho_test_labels.csv').groupby(['customer_id'])
len(dg)

d = []
i = 0
print len(dg)
for k, g in dg:
    i += 1
    if i % 10000 == 0:
        print i / 10000
    g = g.sort_values('measurement_date')
    if len(g) != 2:
        continue
    e = {}
    e['customer_id'] = k
    e['target'] = ''
    for p in products:
        if g[p].tolist() == [0.0, 1.0]:
            e['target'] = e['target'] + ' ' + p
            e[p] = 1
        else:
            e[p] = 0
    e['target'] = e['target'].strip()
    d.append(e)
print i
d = pd.DataFrame(d)

d.head(10)

d[d.target != ''].head(10)

len(d[d.target != ''])

d.to_csv('ho_targets.csv', index=False)
