==> 11_evaluation.ipynb <==
import base64
import datetime
import gc
import json
import pickle
import random
import time
import numpy as np
import pandas as pd
import matplotlib as mpl
import matplotlib.pyplot as plt
from pylab import rcParams
import seaborn as sns
import statsmodels as sm
from pylab import rcParams
from numpy import median
import matplotlib.dates as mdates

from xgboost import XGBClassifier
import sklearn
from sklearn.metrics import confusion_matrix
from sklearn.metrics import roc_auc_score
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import MultinomialNB
from sklearn.tree import DecisionTreeClassifier
from sklearn.grid_search import GridSearchCV
from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier

np.random.seed(1337)
%matplotlib inline
sns.set(font_scale=1.5)
rcParams['figure.figsize'] = 12, 8
sns.set_style('whitegrid')
sns.set_palette(sns.color_palette('muted'))

def pk(true, pred, k):
    pred = pred[:k]
    if k <= len(pred) and pred[k-1] in true:
        return 1.0 * len(set(true) & set(pred)) / len(pred)
    else:
        return 0.0
def apk(true, pred, k):
    if true == '' or pred == '':
        return 0.0
    true = true.split(' ')
    pred = pred.split(' ')
    m = len(true)
    n = len(pred)
    s = 0.0
    for i in range(1, 8):
        s += pk(true, pred, i)
    return 1.0 * s / min(m, 7)

def rank(s):
    return ' '.join(s.sort_values(ascending=False).index[:7])

products = '''ind_recibo_ult1
ind_cco_fin_ult1
ind_nom_pens_ult1
ind_nomina_ult1
ind_cno_fin_ult1
ind_tjcr_fin_ult1
ind_ecue_fin_ult1
ind_dela_fin_ult1
ind_reca_fin_ult1
ind_ctma_fin_ult1
ind_valo_fin_ult1
ind_fond_fin_ult1
ind_deco_fin_ult1
ind_ctop_fin_ult1
ind_ctpp_fin_ult1
ind_plan_fin_ult1
ind_ctju_fin_ult1
ind_deme_fin_ult1
ind_cder_fin_ult1
ind_pres_fin_ult1
ind_hip_fin_ult1
ind_viv_fin_ult1
ind_aval_fin_ult1
ind_ahor_fin_ult1'''.split('\n')

models = {}
for p in products:
    print p
    d = pd.read_csv('tmp/dmy_ho_train_{0}.csv'.format(p))
    del d['customer_id']
    del d['target_name']
    l = d.target.copy()
    del d['target']
    m = XGBClassifier(
                n_estimators=50,
                learning_rate=0.01,
                max_depth=9,
                gamma=0,
                min_child_weight=1,
                scale_pos_weight=5.0,
                nthread=2)
    m.fit(d, l)
    del d
    del l
    gc.collect()
    models[p] = m

pickle.dump(models, open('tmp/models_xgb_ne50_spw5.pickle', 'w'))

models = pickle.load(open('tmp/models_xgb_ne50_spw5.pickle', 'r'))

t = pd.read_csv('tmp/ho_targets.csv')
t.target.fillna('', inplace=True)
t_ids = t[['customer_id']].copy()
del t['customer_id']
t.sum().sort_values(ascending=False)

d = pd.read_csv('tmp/dmy_ho_test.csv')
d = pd.merge(t_ids, d, how='left', on='customer_id')
del d['customer_id']
del d['target_name']
del d['target']

s = pd.DataFrame()
for p in products:
    s[p] = models[p].predict_proba(d)[:,1]
del d
s.sum().sort_values(ascending=False)

s.to_csv('tmp/s_xgb_ne50_spw5.csv', index=False)

s = pd.read_csv('tmp/s_xgb_ne50_spw5.csv')
s['target'] = s.apply(lambda r: rank(r), axis=1)
true = t.target.tolist()
pred = s.target.tolist()
x = 0
for i in range(len(true)):
    x += apk(true[i], pred[i], 7)
print x / len(true)

for p in products[:3]:
    cmx = pd.DataFrame(confusion_matrix(t[p], s[p].apply(lambda x: 1 if x > 0.6 else 0)))
    cmx.index.name='actual'
    cmx.columns.name='predicted'
    print p
    print 'confusion matrix:'
    print cmx
    print '=' * 20    

t[t.target != ''].target.head(20)

s[t.target != ''].target.head(20)

true = t.target.tolist()
pred = ['ind_recibo_ult1 ind_nom_pens_ult1 ind_nomina_ult1 ind_tjcr_fin_ult1 ind_cco_fin_ult1 ind_ecue_fin_ult1 ind_cno_fin_ult1'] * len(true)
x = 0
for i in range(len(true)):
    x += apk(true[i], pred[i], 7)
print x / len(true)
